refactor metrics to adapters
