+++ internal/core/services/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "services",
-    srcs = ["identity_service.go"],
+    srcs = [
+        "identity_service.go",
+        "metrics.go",
+    ],
     importpath = "github.com/sufield/ephemos/internal/core/services",
     visibility = ["//:__subpackages__"],
     deps = [
